requirements.txt
hachoir==3.1.1
Pillow==7.2.0
